passagge/requirements.txt
fastapi
pymongo
python-dotenv
rich